use long to ensure safety

+++ src/main/java/org/agmas/noellesroles/init/RoleShopHandler.java
@@ -115,7 +115,7 @@ public class RoleShopHandler {
 
     public static boolean haveRegistered = false;
     /** 商店内容版本号，客户端 tooltip 索引据此重建。 */
-    public static int shopVersion = 0;
+    public static long shopVersion = 0;
 
     private static List<ShopEntry> createPoisonerShopEntries() {
         var entries = new ArrayList<ShopEntry>();
